@@ -1,9 +1,9 @@
 [package]
 name = "owl_wave"
-version = "0.4.1"
+version = "0.5.0"
 authors = ["Stas Shtin <antisvin@gmail.com>"]
 edition = "2021"
-rust-version = "1.61"
+rust-version = "1.66"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -11,8 +11,8 @@ rust-version = "1.61"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-egui = "0.18.1"
-eframe = { version = "0.18.0", features = ["persistence"] }
+egui = "0.20.1"
+eframe = { version = "0.20.1", features = ["persistence"] }
 serde = { version = "1", features = ["derive"] } # You only need this if you want app persistence
 wavetable = { git = "https://github.com/icsga/Wavetable.git", version = "0.2.0", features = ["use_double_precision"] }
 # default_features = false }
@@ -29,7 +29,7 @@ byte-unit = "4.0.14"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tracing-subscriber = "0.3"
-cpal = { version = "0.13.5" }
+cpal = { version = "0.14.2" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 web-sys = "^0.3.55"
@@ -37,6 +37,7 @@ js-sys = "^0.3"
 console_error_panic_hook = "0.1.7"
 cpal = { version = "0.13.5", features = ["wasm-bindgen"] }
 tracing-wasm = "0.2"
+wasm-bindgen-futures = "0.4"
 
 [features]
 default = []
@@ -46,3 +47,9 @@ jack = ["cpal/jack"]
 
 [profile.release]
 opt-level = 2 # fast and small wasm
+
+# Optimize all dependencies even in debug builds:
+[profile.dev.package."*"]
+opt-level = 2
+
+[patch.crates-io]
